[🚚rename]: PSHeader 파일을 공통 컴포넌트로 이동
